requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
requests==2.31.0
psutil==5.9.5
